One view and one presenter
