+++ k9mail/src/main/java/com/fsck/k9/activity/setup/CheckDirection.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.activity.setup.checksettings;
+package com.fsck.k9.activity.setup;
 
 
 public enum CheckDirection {
