+++ k9mail/src/main/res/layout/account_setup_basics.xml
@@ -22,12 +22,12 @@
         android:id="@+id/email_input_layout"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="0dp"
         app:layout_constraintLeft_toLeftOf="@+id/title"
         android:layout_marginTop="100dp"
         app:layout_constraintTop_toBottomOf="@+id/title"
         android:layout_marginRight="24dp"
-        app:layout_constraintRight_toRightOf="parent">
+        app:layout_constraintRight_toRightOf="parent"
+        android:layout_marginEnd="24dp">
 
         <android.support.design.widget.TextInputEditText
             android:id="@+id/account_email"
@@ -41,13 +41,13 @@
         android:id="@+id/password_input_layout"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="0dp"
         app:layout_constraintLeft_toLeftOf="@+id/title"
         android:layout_marginTop="24dp"
         app:layout_constraintTop_toBottomOf="@+id/email_input_layout"
         android:layout_marginRight="24dp"
         app:passwordToggleEnabled="true"
-        app:layout_constraintRight_toRightOf="parent">
+        app:layout_constraintRight_toRightOf="parent"
+        android:layout_marginEnd="24dp">
 
         <android.support.design.widget.TextInputEditText
             android:id="@+id/account_password"
@@ -57,28 +57,51 @@
             android:hint="password" />
     </android.support.design.widget.TextInputLayout>
 
-    <View
-        android:layout_width="0dp"
-        android:layout_height="56dp"
-        android:background="@color/bottom_bar"
+    <LinearLayout
+        android:id="@+id/button_bar"
         app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginBottom="0dp"
-        android:layout_marginLeft="0dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        android:id="@+id/view2" />
-
-    <TextView
-        android:id="@+id/next"
-        android:layout_width="wrap_content"
+        style="?android:attr/buttonBarStyle"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:text="Next"
-        android:clickable="true"
-        app:layout_constraintTop_toTopOf="@+id/view2"
-        android:layout_marginTop="8dp"
-        android:layout_marginRight="24dp"
-        app:layout_constraintRight_toRightOf="@+id/view2"
-        app:layout_constraintBottom_toBottomOf="@+id/view2"
-        android:layout_marginBottom="8dp" />
+        android:orientation="horizontal"
+        tools:layout_constraintRight_creator="1"
+        android:layout_marginStart="8dp"
+        android:layout_marginEnd="8dp"
+        app:layout_constraintRight_toRightOf="parent"
+        tools:layout_constraintLeft_creator="1"
+        app:layout_constraintLeft_toLeftOf="parent"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp">
+
+        <View
+            android:layout_width="0dp"
+            android:layout_height="1dp"
+            android:layout_weight="1" />
+
+        <Button
+            android:id="@+id/basics_next"
+            style="?android:attr/buttonBarButtonStyle"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:background="@drawable/selectable_item_background"
+            android:text="@string/next_action" />
+    </LinearLayout>
+
+    <View
+        app:layout_constraintBottom_toTopOf="@+id/button_bar"
+        android:id="@+id/divider"
+        android:layout_width="0dp"
+        android:layout_height="1dp"
+        android:background="?android:attr/dividerHorizontal"
+        tools:layout_constraintRight_creator="1"
+        android:layout_marginStart="8dp"
+        android:layout_marginEnd="8dp"
+        app:layout_constraintRight_toRightOf="parent"
+        tools:layout_constraintLeft_creator="1"
+        app:layout_constraintLeft_toLeftOf="parent"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp" />
 
     <Button
         android:id="@+id/manual_setup"
