+++ k9mail/src/main/res/layout/account_setup_incoming.xml
@@ -26,7 +26,7 @@
                 android:textColor="?android:attr/textColorPrimary"
                 android:contentDescription="@string/account_setup_incoming_password_label" />
             <EditText
-                android:id="@+id/account_server"
+                android:id="@+id/incoming_account_server"
                 android:singleLine="true"
                 android:inputType="textUri"
                 android:layout_height="wrap_content"
@@ -38,7 +38,7 @@
                 android:textAppearance="?android:attr/textAppearanceSmall"
                 android:textColor="?android:attr/textColorPrimary" />
             <Spinner
-                android:id="@+id/account_security_type"
+                android:id="@+id/incoming_account_security_type"
                 android:layout_height="wrap_content"
                 android:layout_width="fill_parent"
                 android:contentDescription="@string/account_setup_incoming_security_label" />
@@ -49,7 +49,7 @@
                 android:textAppearance="?android:attr/textAppearanceSmall"
                 android:textColor="?android:attr/textColorPrimary" />
             <EditText
-                android:id="@+id/account_port"
+                android:id="@+id/incoming_account_port"
                 android:singleLine="true"
                 android:inputType="number"
                 android:layout_height="wrap_content"
@@ -62,7 +62,7 @@
                 android:textAppearance="?android:attr/textAppearanceSmall"
                 android:textColor="?android:attr/textColorPrimary" />
             <EditText
-                android:id="@+id/account_username"
+                android:id="@+id/incoming_account_username"
                 android:singleLine="true"
                 android:inputType="textEmailAddress"
                 android:layout_height="wrap_content"
@@ -76,7 +76,7 @@
                 android:textAppearance="?android:attr/textAppearanceSmall"
                 android:textColor="?android:attr/textColorPrimary" />
             <Spinner
-                android:id="@+id/account_auth_type"
+                android:id="@+id/incoming_account_auth_type"
                 android:layout_height="wrap_content"
                 android:layout_width="fill_parent"
                 android:contentDescription="@string/account_setup_incoming_auth_type_label" />
@@ -88,7 +88,7 @@
                 android:textAppearance="?android:attr/textAppearanceSmall"
                 android:textColor="?android:attr/textColorPrimary" />
             <EditText
-                android:id="@+id/account_password"
+                android:id="@+id/incoming_account_password"
                 android:inputType="textPassword"
                 android:singleLine="true"
                 android:layout_height="wrap_content"
@@ -102,7 +102,7 @@
                 android:textColor="?android:attr/textColorPrimary"
                 android:visibility="gone" />
             <com.fsck.k9.view.ClientCertificateSpinner
-                android:id="@+id/account_client_certificate_spinner"
+                android:id="@+id/incoming_account_client_certificate_spinner"
                 android:layout_height="wrap_content"
                 android:layout_width="fill_parent"
                 android:visibility="gone" />
@@ -249,6 +249,31 @@
                 android:layout_weight="1" />
         </LinearLayout>
     </ScrollView>
-    <include layout="@layout/wizard_next" />
+    <View
+        android:id="@+id/divider"
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:background="?android:attr/dividerHorizontal" />
+
+    <LinearLayout
+        style="?android:attr/buttonBarStyle"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal" >
+
+        <View
+            android:layout_width="0dp"
+            android:layout_height="1dp"
+            android:layout_weight="1" />
+
+        <Button
+            android:id="@+id/incoming_next"
+            style="?android:attr/buttonBarButtonStyle"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:background="@drawable/selectable_item_background"
+            android:text="@string/next_action" />
+    </LinearLayout>
 </LinearLayout>
