+++ k9mail/src/main/res/layout/account_setup_names.xml
@@ -5,38 +5,38 @@
     android:layout_height="fill_parent"
     android:orientation="vertical">
 
-<ScrollView
-    android:layout_width="fill_parent"
-    android:layout_height="0dp"
-    android:layout_weight="1"
-    android:padding="6dip"
-    android:fadingEdge="none"
-    android:scrollbarStyle="outsideInset">
-    <LinearLayout
+    <ScrollView
         android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:layout_gravity="center_horizontal|center_vertical"
-        android:orientation="vertical">
-    <EditText
-        android:id="@+id/account_description"
-        android:singleLine="true"
-        android:layout_height="wrap_content"
-        android:layout_width="fill_parent"
-        android:hint="@string/account_setup_names_account_name_label"
-        android:contentDescription="@string/account_setup_names_account_name_label" />
-    <EditText
-        android:id="@+id/account_name"
-        android:singleLine="true"
-        android:inputType="textPersonName"
-        android:layout_height="wrap_content"
-        android:layout_width="fill_parent"
-        android:hint="@string/account_setup_names_user_name_label"
-        android:contentDescription="@string/account_setup_names_user_name_label" />
-    <View
-        android:layout_height="0dip"
-        android:layout_width="fill_parent"
-        android:layout_weight="1" />
-</LinearLayout>
-</ScrollView>
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:padding="6dip"
+        android:fadingEdge="none"
+        android:scrollbarStyle="outsideInset">
+        <LinearLayout
+            android:layout_width="fill_parent"
+            android:layout_height="fill_parent"
+            android:layout_gravity="center_horizontal|center_vertical"
+            android:orientation="vertical">
+            <EditText
+                android:id="@+id/account_description"
+                android:singleLine="true"
+                android:layout_height="wrap_content"
+                android:layout_width="fill_parent"
+                android:hint="@string/account_setup_names_account_name_label"
+                android:contentDescription="@string/account_setup_names_account_name_label" />
+            <EditText
+                android:id="@+id/account_name"
+                android:singleLine="true"
+                android:inputType="textPersonName"
+                android:layout_height="wrap_content"
+                android:layout_width="fill_parent"
+                android:hint="@string/account_setup_names_user_name_label"
+                android:contentDescription="@string/account_setup_names_user_name_label" />
+            <View
+                android:layout_height="0dip"
+                android:layout_width="fill_parent"
+                android:layout_weight="1" />
+        </LinearLayout>
+    </ScrollView>
     <include layout="@layout/wizard_done" />
 </LinearLayout>
