+++ k9mail/src/main/res/layout/account_setup_outgoing.xml
@@ -26,7 +26,7 @@
                 android:textColor="?android:attr/textColorPrimary" />
 
             <EditText
-                android:id="@+id/account_server"
+                android:id="@+id/outgoing_account_server"
                 android:singleLine="true"
                 android:inputType="textUri"
                 android:layout_height="wrap_content"
@@ -41,7 +41,7 @@
                 android:textColor="?android:attr/textColorPrimary" />
 
             <Spinner
-                android:id="@+id/account_security_type"
+                android:id="@+id/outgoing_account_security_type"
                 android:layout_height="wrap_content"
                 android:layout_width="fill_parent"
                 android:contentDescription="@string/account_setup_outgoing_security_label" />
@@ -54,7 +54,7 @@
                 android:textColor="?android:attr/textColorPrimary" />
 
             <EditText
-                android:id="@+id/account_port"
+                android:id="@+id/outgoing_account_port"
                 android:singleLine="true"
                 android:inputType="number"
                 android:layout_height="wrap_content"
@@ -82,7 +82,7 @@
                     android:textColor="?android:attr/textColorPrimary" />
 
                 <EditText
-                    android:id="@+id/account_username"
+                    android:id="@+id/outgoing_account_username"
                     android:singleLine="true"
                     android:inputType="textEmailAddress"
                     android:layout_height="wrap_content"
@@ -97,7 +97,7 @@
                     android:textColor="?android:attr/textColorPrimary" />
 
                 <Spinner
-                    android:id="@+id/account_auth_type"
+                    android:id="@+id/outgoing_account_auth_type"
                     android:layout_height="wrap_content"
                     android:layout_width="fill_parent"
                     android:contentDescription="@string/account_setup_outgoing_authentication_label" />
@@ -111,7 +111,7 @@
                     android:textColor="?android:attr/textColorPrimary" />
 
                 <EditText
-                    android:id="@+id/account_password"
+                    android:id="@+id/outgoing_account_password"
                     android:singleLine="true"
                     android:inputType="textPassword"
                     android:layout_height="wrap_content"
@@ -128,7 +128,7 @@
                     android:visibility="gone" />
 
                 <com.fsck.k9.view.ClientCertificateSpinner
-                    android:id="@+id/account_client_certificate_spinner"
+                    android:id="@+id/outgoing_account_client_certificate_spinner"
                     android:layout_height="wrap_content"
                     android:layout_width="fill_parent"
                     android:visibility="gone" />
@@ -141,6 +141,31 @@
         </LinearLayout>
     </ScrollView>
 
-    <include layout="@layout/wizard_next" />
+    <View
+        android:id="@+id/divider"
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:background="?android:attr/dividerHorizontal" />
+
+    <LinearLayout
+        style="?android:attr/buttonBarStyle"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal" >
+
+        <View
+            android:layout_width="0dp"
+            android:layout_height="1dp"
+            android:layout_weight="1" />
+
+        <Button
+            android:id="@+id/outgoing_next"
+            style="?android:attr/buttonBarButtonStyle"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:background="@drawable/selectable_item_background"
+            android:text="@string/next_action" />
+    </LinearLayout>
 
 </LinearLayout>
